Improved profile picture renderer
