+++ lib/connection/encryption/signatures.dart
@@ -12,18 +12,13 @@ KeyPair generateSignatureKeyPair([Sodium? sd]) {
 String signMessage(SecureKey privateKey, String message, [Sodium? sd]) {
   final Sodium sodium = sd ?? sodiumLib;
   final plainTextBytes = message.toCharArray().unsignedView();
-  final signed = sodium.crypto.sign
-      .detached(message: plainTextBytes, secretKey: privateKey);
+  final signed = sodium.crypto.sign.detached(message: plainTextBytes, secretKey: privateKey);
   return base64Encode(signed);
 }
 
 /// message = message to check, signature = "encrypted" signature, publicKey = public key of the sender
-bool checkSignature(String signature, Uint8List publicKey, String message,
-    [Sodium? sd]) {
+bool checkSignature(String signature, Uint8List publicKey, String message, [Sodium? sd]) {
   final Sodium sodium = sd ?? sodiumLib;
   final plainTextBytes = message.toCharArray().unsignedView();
-  return sodium.crypto.sign.verifyDetached(
-      signature: base64Decode(signature),
-      message: plainTextBytes,
-      publicKey: publicKey);
+  return sodium.crypto.sign.verifyDetached(signature: base64Decode(signature), message: plainTextBytes, publicKey: publicKey);
 }
