+++ lib/theme/ui/dialogs/profile_picture_window.dart
@@ -1,3 +1,4 @@
+import 'dart:io';
 import 'dart:ui' as ui;
 
 import 'package:chat_interface/controller/account/profile_picture_helper.dart';
@@ -8,6 +9,9 @@ import 'package:chat_interface/util/vertical_spacing.dart';
 import 'package:file_selector/file_selector.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:screenshot/screenshot.dart';
+import 'package:path/path.dart' as path;
 
 class ProfilePictureWindow extends StatefulWidget {
   final XFile file;
@@ -138,7 +142,27 @@ class _ProfilePictureWindowState extends State<ProfilePictureWindow> {
                 onTap: () async {
                   if (uploading.value) return;
                   uploading.value = true;
-                  await ProfilePictureHelper.uploadProfilePicture(widget.file, ProfilePictureData(scaleFactor.value, moveX.value, moveY.value));
+
+                  final screenshotController = ScreenshotController();
+                  final scale = scaleFactor.value * (300 / 500);
+
+                  final image = await screenshotController.captureFromWidget(
+                    SizedBox(
+                      width: 500,
+                      height: 500,
+                      child: RawImage(
+                        fit: BoxFit.none,
+                        scale: scale,
+                        image: _image.value!,
+                        alignment: Alignment(moveX.value, moveY.value),
+                      ),
+                    ),
+                  );
+                  final currentFile = File(widget.file.path);
+                  final cutFile = File(path.join(currentFile.parent.path, ".cut-${widget.file.name}"));
+                  await cutFile.writeAsBytes(image);
+                  await ProfilePictureHelper.uploadProfilePicture(cutFile, widget.file.name);
+                  await cutFile.delete();
                   uploading.value = false;
                   Get.back();
                 },
