+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   archive: '>=3.4.10 <4.0.0'
   window_manager: '>=0.3.8 <1.0.0'
   permission_handler: '>=11.3.1 <12.0.0'
+  screenshot: ^2.3.0
 
 dev_dependencies:
   flutter_test:
